# api/cloud_storage/v1/user_router.pb.go
// Code generated by https://github.com/go-dev-frame/sponge, DO NOT EDIT.

package v1

import (
	"context"
	"errors"
	"strings"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	"github.com/go-dev-frame/sponge/pkg/errcode"
	"github.com/go-dev-frame/sponge/pkg/gin/middleware"
)

type UserLogicer interface {
	UserLogin(ctx context.Context, req *LoginRequest) (*LoginReply, error)
	UserDetail(ctx context.Context, req *UserDetailRequest) (*UserDetailReply, error)
	MailCodeSendRegister(ctx context.Context, req *MailCodeSendRequest) (*MailCodeSendReply, error)
	UserRegister(ctx context.Context, req *UserRegisterRequest) (*UserRegisterReply, error)
	RefreshAuthorization(ctx context.Context, req *RefreshAuthorizationRequest) (*RefreshAuthorizationReply, error)
}

type UserOption func(*userOptions)

type userOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *userOptions) apply(opts ...UserOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithUserHTTPResponse() UserOption {
	return func(o *userOptions) {
		o.isFromRPC = false
	}
}

func WithUserRPCResponse() UserOption {
	return func(o *userOptions) {
		o.isFromRPC = true
	}
}

func WithUserResponser(responser errcode.Responser) UserOption {
	return func(o *userOptions) {
		o.responser = responser
	}
}

func WithUserLogger(zapLog *zap.Logger) UserOption {
	return func(o *userOptions) {
		o.zapLog = zapLog
	}
}

func WithUserErrorToHTTPCode(e ...*errcode.Error) UserOption {
	return func(o *userOptions) {
		o.httpErrors = e
	}
}

func WithUserRPCStatusToHTTPCode(s ...*errcode.RPCStatus) UserOption {
	return func(o *userOptions) {
		o.rpcStatus = s
	}
}

func WithUserWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) UserOption {
	return func(o *userOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterUserRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic UserLogicer,
	opts ...UserOption) {

	o := &userOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &userRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type userRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                UserLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *userRouter) register() {
	r.iRouter.Handle("POST", "/user/login", r.withMiddleware("POST", "/user/login", r.UserLogin_0)...)
	r.iRouter.Handle("POST", "/user/detail", r.withMiddleware("POST", "/user/detail", r.UserDetail_0)...)
	r.iRouter.Handle("POST", "/mail/code/send/register", r.withMiddleware("POST", "/mail/code/send/register", r.MailCodeSendRegister_0)...)
	r.iRouter.Handle("POST", "/user/register", r.withMiddleware("POST", "/user/register", r.UserRegister_0)...)
	r.iRouter.Handle("POST", "/refresh/authorization", r.withMiddleware("POST", "/refresh/authorization", r.RefreshAuthorization_0)...)

}

func (r *userRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

func (r *userRouter) UserLogin_0(c *gin.Context) {
	req := &LoginRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UserLogin(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *userRouter) UserDetail_0(c *gin.Context) {
	req := &UserDetailRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UserDetail(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *userRouter) MailCodeSendRegister_0(c *gin.Context) {
	req := &MailCodeSendRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.MailCodeSendRegister(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *userRouter) UserRegister_0(c *gin.Context) {
	req := &UserRegisterRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UserRegister(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *userRouter) RefreshAuthorization_0(c *gin.Context) {
	req := &RefreshAuthorizationRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.RefreshAuthorization(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}