# docs/apis.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/cloud_storage/v1/chunk.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/mail/code/send/register": {
      "post": {
        "operationId": "user_MailCodeSendRegister",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MailCodeSendRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/MailCodeSendHTTPResponse"
            }
          }
        },
        "summary": "验证码发送",
        "tags": [
          "user"
        ]
      }
    },
    "/refresh/authorization": {
      "post": {
        "operationId": "user_RefreshAuthorization",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RefreshAuthorizationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/RefreshAuthorizationHTTPResponse"
            }
          }
        },
        "summary": "刷新Authorization",
        "tags": [
          "user"
        ]
      }
    },
    "/share/basic/create": {
      "post": {
        "operationId": "share_ShareBasicCreate",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ShareBasicCreateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ShareBasicCreateHTTPResponse"
            }
          }
        },
        "summary": "创建分享记录",
        "tags": [
          "share"
        ]
      }
    },
    "/share/basic/detail": {
      "get": {
        "operationId": "share_ShareBasicDetail",
        "parameters": [
          {
            "in": "query",
            "name": "identity",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ShareBasicDetailHTTPResponse"
            }
          }
        },
        "summary": "获取资源详情",
        "tags": [
          "share"
        ]
      }
    },
    "/share/basic/save": {
      "post": {
        "operationId": "share_ShareBasicSave",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ShareBasicSaveRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ShareBasicSaveHTTPResponse"
            }
          }
        },
        "summary": "资源保存",
        "tags": [
          "share"
        ]
      }
    },
    "/user/detail": {
      "post": {
        "operationId": "user_UserDetail",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserDetailRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/UserDetailHTTPResponse"
            }
          }
        },
        "summary": "用户详情",
        "tags": [
          "user"
        ]
      }
    },
...
    "/user/login": {
      "post": {
        "operationId": "user_UserLogin",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/LoginHTTPResponse"
            }
          }
        },
        "summary": "用户登录",
        "tags": [
          "user"
        ]
      }
    },
    "/user/register": {
      "post": {
        "operationId": "user_UserRegister",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/UserRegisterHTTPResponse"
            }
          }
        },
        "summary": "用户注册",
        "tags": [
          "user"
        ]
      }
    },
...
  },
  "definitions": {
...
    "LoginHTTPResponse": {
      "properties": {
        "code": {
          "description": "Business status code",
          "format": "int32",
          "type": "integer"
        },
        "data": {
          "$ref": "#/definitions/LoginReply"
        },
        "msg": {
          "description": "Response message description",
          "type": "string"
        }
      },
      "type": "object"
    },
    "LoginReply": {
      "properties": {
        "refreshToken": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "LoginRequest": {
      "properties": {
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "MailCodeSendHTTPResponse": {
      "properties": {
        "code": {
          "description": "Business status code",
          "format": "int32",
          "type": "integer"
        },
        "data": {
          "$ref": "#/definitions/MailCodeSendReply"
        },
        "msg": {
          "description": "Response message description",
          "type": "string"
        }
      },
      "type": "object"
    },
    "MailCodeSendReply": {
      "type": "object"
    },
    "MailCodeSendRequest": {
      "properties": {
        "email": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "RefreshAuthorizationHTTPResponse": {
      "properties": {
        "code": {
          "description": "Business status code",
          "format": "int32",
          "type": "integer"
        },
        "data": {
          "$ref": "#/definitions/RefreshAuthorizationReply"
        },
        "msg": {
          "description": "Response message description",
          "type": "string"
        }
      },
      "type": "object"
    },
    "RefreshAuthorizationReply": {
      "properties": {
        "refreshToken": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "RefreshAuthorizationRequest": {
      "type": "object"
    },
    "ShareBasicCreateHTTPResponse": {
      "properties": {
        "code": {
          "description": "Business status code",
          "format": "int32",
          "type": "integer"
        },
        "data": {
          "$ref": "#/definitions/ShareBasicCreateReply"
        },
        "msg": {
          "description": "Response message description",
          "type": "string"
        }
      },
      "type": "object"
    },
    "ShareBasicCreateReply": {
      "properties": {
        "identity": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "ShareBasicCreateRequest": {
      "properties": {
        "expiredTime": {
          "format": "int32",
          "type": "integer"
        },
        "userRepositoryIdentity": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "ShareBasicDetailHTTPResponse": {
      "properties": {
        "code": {
          "description": "Business status code",
          "format": "int32",
          "type": "integer"
        },
        "data": {
          "$ref": "#/definitions/ShareBasicDetailReply"
        },
        "msg": {
          "description": "Response message description",
          "type": "string"
        }
      },
      "type": "object"
    },
    "ShareBasicDetailReply": {
      "properties": {
        "ext": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "repositoryIdentity": {
          "type": "string"
        },
        "size": {
          "format": "int64",
          "type": "integer"
        }
      },
      "type": "object"
    },
    "ShareBasicSaveHTTPResponse": {
      "properties": {
        "code": {
          "description": "Business status code",
          "format": "int32",
          "type": "integer"
        },
        "data": {
          "$ref": "#/definitions/ShareBasicSaveReply"
        },
        "msg": {
          "description": "Response message description",
          "type": "string"
        }
      },
      "type": "object"
    },
    "ShareBasicSaveReply": {
      "properties": {
        "identity": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "ShareBasicSaveRequest": {
      "properties": {
        "parentId": {
          "format": "int64",
          "type": "integer"
        },
        "repositoryIdentity": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "UserDetailHTTPResponse": {
      "properties": {
        "code": {
          "description": "Business status code",
          "format": "int32",
          "type": "integer"
        },
        "data": {
          "$ref": "#/definitions/UserDetailReply"
        },
        "msg": {
          "description": "Response message description",
          "type": "string"
        }
      },
      "type": "object"
    },
    "UserDetailReply": {
      "properties": {
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "UserDetailRequest": {
      "properties": {
        "identity": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "UserRegisterHTTPResponse": {
      "properties": {
        "code": {
          "description": "Business status code",
          "format": "int32",
          "type": "integer"
        },
        "data": {
          "$ref": "#/definitions/UserRegisterReply"
        },
        "msg": {
          "description": "Response message description",
          "type": "string"
        }
      },
      "type": "object"
    },
    "UserRegisterReply": {
      "type": "object"
    },
    "UserRegisterRequest": {
      "properties": {
        "code": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "type": "object"
    },
...
  }
}